Updated searcher.js to cater for Windows paths
Windows paths contain semicolon (e.g. C:\Windows), while searcher.js has to split at semicolon, since rg uses it to separate path, line, column and code.

If the first array element has a length of 1, it can safely be assumed that it's a Windows path, so concatenate the first and second elements. (In Linux, the minimum path will be '/a' so the length is at least 2).